update PostService to use token, update app.js /posts to use token again, change reference FeedPageView to FeedView, remove FeedPageView
